+++ Authenticator/Source/InfoListViewController.swift
@@ -2,7 +2,7 @@
 //  InfoListViewController.swift
 //  Authenticator
 //
-//  Copyright (c) 2017 Authenticator authors
+//  Copyright (c) 2017-2018 Authenticator authors
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -37,6 +37,7 @@ final class InfoListViewController: UITableViewController {
         super.init(nibName: nil, bundle: nil)
     }
 
+    @available(*, unavailable)
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -110,9 +111,9 @@ class InfoListCell: UITableViewCell {
         return paragraphStyle
     }()
 
-    let titleLabel = UILabel()
-    let descriptionLabel = UILabel()
-    var customConstraints: [NSLayoutConstraint]?
+    private let titleLabel = UILabel()
+    private let descriptionLabel = UILabel()
+    private var customConstraints: [NSLayoutConstraint]?
 
     // MARK: Initialization