+++ Authenticator/Source/Root.swift
@@ -2,7 +2,7 @@
 //  Root.swift
 //  Authenticator
 //
-//  Copyright (c) 2015-2017 Authenticator authors
+//  Copyright (c) 2015-2018 Authenticator authors
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -36,9 +36,9 @@ struct Root: Component {
         case scanner(TokenScanner)
         case entryForm(TokenEntryForm)
         case editForm(TokenEditForm)
-        case info(InfoList, Info?)
+        case menu(Menu)
 
-        var viewModel: RootViewModel.ModalViewModel {
+        func viewModel(digitGroupSize: Int) -> RootViewModel.ModalViewModel {
             switch self {
             case .none:
                 return .none
@@ -48,8 +48,8 @@ struct Root: Component {
                 return .entryForm(form.viewModel)
             case .editForm(let form):
                 return .editForm(form.viewModel)
-            case let .info(infoList, info):
-                return .info(infoList.viewModel, info?.viewModel)
+            case let .menu(menu):
+                return .menu(menu.viewModel(digitGroupSize: digitGroupSize))
             }
         }
     }
@@ -66,11 +66,15 @@ struct Root: Component {
 extension Root {
     typealias ViewModel = RootViewModel
 
-    func viewModel(with persistentTokens: [PersistentToken], at displayTime: DisplayTime) -> (viewModel: ViewModel, nextRefreshTime: Date) {
-        let (tokenListViewModel, nextRefreshTime) = tokenList.viewModel(with: persistentTokens, at: displayTime)
+    func viewModel(with persistentTokens: [PersistentToken], at displayTime: DisplayTime, digitGroupSize: Int) -> (viewModel: ViewModel, nextRefreshTime: Date) {
+        let (tokenListViewModel, nextRefreshTime) = tokenList.viewModel(
+            with: persistentTokens,
+            at: displayTime,
+            digitGroupSize: digitGroupSize
+        )
         let viewModel = ViewModel(
             tokenList: tokenListViewModel,
-            modal: modal.viewModel
+            modal: modal.viewModel(digitGroupSize: digitGroupSize)
         )
         return (viewModel: viewModel, nextRefreshTime: nextRefreshTime)
     }
@@ -87,7 +91,9 @@ extension Root {
 
         case infoListEffect(InfoList.Effect)
         case infoEffect(Info.Effect)
+        case displayOptionsEffect(DisplayOptions.Effect)
         case dismissInfo
+        case dismissDisplayOptions
 
         case addTokenFromURL(Token)
     }
@@ -125,6 +131,7 @@ extension Root {
         case showSuccessMessage(String)
         case showApplicationSettings
         case openURL(URL)
+        case setDigitGroupSize(Int)
     }
 
     mutating func update(with action: Action) throws -> Effect? {
@@ -160,8 +167,19 @@ extension Root {
             case .infoEffect(let effect):
                 return handleInfoEffect(effect)
 
+            case .displayOptionsEffect(let effect):
+                return handleDisplayOptionsEffect(effect)
+
             case .dismissInfo:
-                try modal.dismissInfo()
+                try modal.withMenu { menu in
+                    try menu.dismissInfo()
+                }
+                return nil
+
+            case .dismissDisplayOptions:
+                try modal.withMenu { menu in
+                    try menu.dismissDisplayOptions()
+                }
                 return nil
 
             case .addTokenFromURL(let token):
@@ -232,14 +250,14 @@ extension Root {
 
         case .showBackupInfo:
             do {
-                modal = .info(InfoList(), try Info.backupInfo())
+                modal = .menu(Menu(info: try Info.backupInfo()))
                 return nil
             } catch {
                 return .showErrorMessage("Failed to load backup info.")
             }
 
         case .showInfo:
-            modal = .info(InfoList(), nil)
+            modal = .menu(Menu())
             return nil
         }
     }
@@ -307,6 +325,12 @@ extension Root {
 
     private mutating func handleInfoListEffect(_ effect: InfoList.Effect) throws -> Effect? {
         switch effect {
+        case .showDisplayOptions:
+            try modal.withMenu { menu in
+                try menu.showDisplayOptions()
+            }
+            return nil
+
         case .showBackupInfo:
             let backupInfo: Info
             do {
@@ -314,7 +338,9 @@ extension Root {
             } catch {
                 return .showErrorMessage("Failed to load backup info.")
             }
-            try modal.setInfo(backupInfo)
+            try modal.withMenu { menu in
+                try menu.showInfo(backupInfo)
+            }
             return nil
 
         case .showLicenseInfo:
@@ -324,7 +350,9 @@ extension Root {
             } catch {
                 return .showErrorMessage("Failed to load acknowledgements.")
             }
-            try modal.setInfo(licenseInfo)
+            try modal.withMenu { menu in
+                try menu.showInfo(licenseInfo)
+            }
             return nil
 
         case .done:
@@ -341,6 +369,16 @@ extension Root {
             return nil
         case let .openURL(url):
             return .openURL(url)
+        }
+    }
+
+    private mutating func handleDisplayOptionsEffect(_ effect: DisplayOptions.Effect) -> Effect? {
+        switch effect {
+        case .done:
+            modal = .none
+            return nil
+        case let .setDigitGroupSize(digitGroupSize):
+            return .setDigitGroupSize(digitGroupSize)
         }
     }
 }
@@ -378,17 +416,12 @@ private extension Root.Modal {
         return result
     }
 
-    mutating func setInfo(_ info: Info) throws {
-        guard case .info(let infoList, .none) = self else {
-            throw Error(expectedType: InfoList.self, actualState: self)
+    mutating func withMenu<ResultType>(_ body: (inout Menu) throws -> ResultType) throws -> ResultType {
+        guard case .menu(var menu) = self else {
+            throw Error(expectedType: Menu.self, actualState: self)
         }
-        self = .info(infoList, info)
-    }
-
-    mutating func dismissInfo() throws {
-        guard case .info(let infoList, .some) = self else {
-            throw Error(expectedType: Info.self, actualState: self)
-        }
-        self = .info(infoList, nil)
+        let result = try body(&menu)
+        self = .menu(menu)
+        return result
     }
 }