+++ Authenticator/Source/TokenFormViewController.swift
@@ -2,7 +2,7 @@
 //  TokenFormViewController.swift
 //  Authenticator
 //
-//  Copyright (c) 2015-2017 Authenticator authors
+//  Copyright (c) 2015-2018 Authenticator authors
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -55,6 +55,7 @@ final class TokenFormViewController<Form: TableViewModelRepresentable>: UITableV
         super.init(style: .grouped)
     }
 
+    @available(*, unavailable)
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
@@ -155,9 +156,9 @@ final class TokenFormViewController<Form: TableViewModelRepresentable>: UITableV
     // MARK: - UITableViewDelegate
 
     override func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        // An apparent rendering error can occur when the form is scrolled programaticallty, causing a cell scrolled off
-        // of the screen to appear with a black background when scrolled back onto the screen. Setting the background
-        // color of the cell to the table view's background color, instead of to clearColor(), fixes the issue.
+        // An apparent rendering error can occur when the table view is scrolled programmatically, causing a cell
+        // scrolled off of the screen to appear with a black background when scrolled back onto the screen. Setting the
+        // background color of the cell to the table view's background color, instead of to `.clear`, fixes the issue.
         cell.backgroundColor = .otpBackgroundColor
         cell.selectionStyle = .none